+++ src/main/java/com/imeira/account/transaction/repository/TransactionRepository.java
@@ -4,7 +4,9 @@ import com.imeira.account.transaction.domain.Transaction;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface TransactionRepository extends MongoRepository<Transaction, Long> {
+public interface TransactionRepository extends MongoRepository<Transaction, BigInteger> {
 
 }
